+++ src/main/java/nl/dtls/fairdatapoint/api/controller/index/AdminController.java
@@ -24,6 +24,7 @@ package nl.dtls.fairdatapoint.api.controller.index;
 
 import io.swagger.annotations.ApiOperation;
 import lombok.extern.log4j.Log4j2;
+import nl.dtls.fairdatapoint.api.dto.index.ping.PingDTO;
 import nl.dtls.fairdatapoint.entity.index.event.Event;
 import nl.dtls.fairdatapoint.service.UtilityService;
 import nl.dtls.fairdatapoint.service.index.event.EventService;
@@ -34,6 +35,7 @@ import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpServletRequest;
+import javax.validation.Valid;
 import java.util.UUID;
 
 @Log4j2
@@ -54,9 +56,20 @@ public class AdminController {
     @PostMapping("/trigger")
     @PreAuthorize("hasRole('ADMIN')")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void triggerMetadataRetrieve(@RequestParam(required = false) String clientUrl, HttpServletRequest request) {
+    public void triggerMetadataRetrieve(@RequestBody @Valid PingDTO reqDto, HttpServletRequest request) {
         log.info("Received ping from {}", utilityService.getRemoteAddr(request));
-        final Event event = eventService.acceptAdminTrigger(request, clientUrl);
+        final Event event = eventService.acceptAdminTrigger(request, reqDto);
+        webhookService.triggerWebhooks(event);
+        eventService.triggerMetadataRetrieval(event);
+    }
+
+    @ApiOperation(value = "trigger-all", hidden = true)
+    @PostMapping("/trigger-all")
+    @PreAuthorize("hasRole('ADMIN')")
+    @ResponseStatus(HttpStatus.NO_CONTENT)
+    public void triggerMetadataRetrieveAll(HttpServletRequest request) {
+        log.info("Received ping from {}", utilityService.getRemoteAddr(request));
+        final Event event = eventService.acceptAdminTriggerAll(request);
         webhookService.triggerWebhooks(event);
         eventService.triggerMetadataRetrieval(event);
     }